requirements.txt
Flask==1.0.2
message_listener==0.1.2
iot_message==0.1.0
python-i18n==0.3.6
